bump and simplify the store-data-on-cell
