+++ examples/store-data-on-cell/ccc-client.ts
@@ -1,17 +1,21 @@
-import offCKB, { Network } from "./offckb.config";
 import { ccc, CellDepInfoLike, KnownScript, Script } from "@ckb-ccc/core";
+import systemScripts from "./system-scripts.json";
+
+export type Network = 'devnet' | 'testnet' | 'mainnet';
+
+export type ScriptInfo = Pick<Script, "codeHash" | "hashType"> & { cellDeps: CellDepInfoLike[] };
 
 export const DEVNET_SCRIPTS: Record<
   string,
-  Pick<Script, "codeHash" | "hashType"> & { cellDeps: CellDepInfoLike[] }
+  ScriptInfo
 > = {
   [KnownScript.Secp256k1Blake160]:
-    offCKB.systemScripts.secp256k1_blake160_sighash_all!.script,
+    systemScripts["devnet"].secp256k1_blake160_sighash_all!.script as ScriptInfo,
   [KnownScript.Secp256k1Multisig]:
-    offCKB.systemScripts.secp256k1_blake160_multisig_all!.script,
-  [KnownScript.AnyoneCanPay]: offCKB.systemScripts.anyone_can_pay!.script,
-  [KnownScript.OmniLock]: offCKB.systemScripts.omnilock!.script,
-  [KnownScript.XUdt]: offCKB.systemScripts.xudt!.script,
+    systemScripts["devnet"].secp256k1_blake160_multisig_all!.script as ScriptInfo,
+  [KnownScript.AnyoneCanPay]: systemScripts["devnet"].anyone_can_pay!.script as ScriptInfo,
+  [KnownScript.OmniLock]: systemScripts["devnet"].omnilock!.script as ScriptInfo,
+  [KnownScript.XUdt]: systemScripts["devnet"].xudt!.script as ScriptInfo,
 };
 
 export function buildCccClient(network: Network) {
@@ -21,11 +25,23 @@ export function buildCccClient(network: Network) {
       : network === "testnet"
       ? new ccc.ClientPublicTestnet()
       : new ccc.ClientPublicTestnet({
-          url: offCKB.rpcUrl,
+          url: "http://localhost:28114", // the default offckb devnet proxy rpc url
           scripts: DEVNET_SCRIPTS as any,
         });
 
   return client;
 }
 
-export const cccClient = buildCccClient(offCKB.currentNetwork);
+export function readEnvNetwork(): Network {
+  const network = process.env.NETWORK;
+  const defaultNetwork = 'testnet';
+  if (!network) return defaultNetwork;
+
+  if (!['devnet', 'testnet', 'mainnet'].includes(network)) {
+    return defaultNetwork;
+  }
+
+  return network as Network;
+}
+
+export const cccClient = buildCccClient(readEnvNetwork());
