--- examples/store-data-on-cell/offckb/my-scripts.json
@@ -1,5 +1,0 @@
-{
-  "devnet": {},
-  "testnet": {},
-  "mainnet": {}
-}
